(edit) create the separate component for 404 error and pass the statusCode prop to ErrorPage
