+++ src/pages/_error.js
@@ -1,66 +1,17 @@
-import { useEffect } from 'react'
-import { useRouter } from 'next/router'
-import { Box, Heading, Paragraph } from 'grommet'
-import { useResponsive } from 'hooks/useResponsive'
-import { Button } from 'components/shared/Button'
 import { FixedContainer } from 'components/shared/FixedContainer'
-import { Row } from 'components/shared/Row'
+import { Custom404 } from 'components/Error/Custom404'
 
-export const ErrorPage = () => {
-  const router = useRouter()
-  const { setResponsive } = useResponsive()
+export const ErrorPage = ({ statusCode }) => {
+  return <FixedContainer>{statusCode === 404 && <Custom404 />}</FixedContainer>
+}
 
-  useEffect(() => {
-    const forceRefresh = (url) => {
-      window.location = url
-    }
-    router.events.on('routeChangeStart', forceRefresh)
-  })
+ErrorPage.getInitialProps = async ({ res, err }) => {
+  // eslint-disable-next-line no-nested-ternary
+  const statusCode = res ? res.statusCode : err ? err.statusCode : 404
 
-  const goBack = () => {
-    router.back()
+  return {
+    statusCode
   }
-
-  return (
-    <FixedContainer>
-      <Row
-        direction={setResponsive('column', 'column', 'row')}
-        gap="xlarge"
-        justify="center"
-        margin={{ top: setResponsive('none', 'none', 'basex15') }}
-      >
-        <Box
-          align={setResponsive('center', 'center', 'start')}
-          margin={{ top: 'basex8' }}
-        >
-          <Heading level={1} size="small">
-            The page you are looking for isn’t expressed.
-          </Heading>
-          <Paragraph>Please try again later..</Paragraph>
-          <Button
-            label="Go Back"
-            margin={{ top: 'medium' }}
-            onClick={goBack}
-            primary
-            responsive
-          />
-        </Box>
-        <Box
-          aria-hidden
-          background={{
-            image: "url('/illustration-reward-poster.svg')",
-            position: 'center',
-            repeat: 'no-repeat',
-            size: 'contain'
-          }}
-          alignSelf={setResponsive('center', 'center', 'start')}
-          // to preserve the height of SVG image
-          height={setResponsive('250px', '350px')}
-          width={setResponsive('250px', '350px')}
-        />
-      </Row>
-    </FixedContainer>
-  )
 }
 
 export default ErrorPage
